[VMCB, vIRQ] Moved guest_virqmap into vmcb
