+++ include/core/virq.h
@@ -9,13 +9,13 @@
 
 struct virq {
     struct vgic_status vgic_status;
-    struct guest_virqmap guest_virqmap;
+    struct guest_virqmap *guest_virqmap;
 };
 
 void virq_setup();
-void virq_create();
+void virq_create(struct virq *virq, vmid_t vmid);
 
-hvmm_status_t virq_init(struct virq *virq);
+hvmm_status_t virq_init(struct virq *virq, vmid_t vmid);
 hvmm_status_t virq_save(struct virq *virq);
 hvmm_status_t virq_restore(struct virq *virq, vmid_t vmid);
